refactor: firebase datasource 구조 분리(stats, datasource)
